+++ RELEASE_NOTES.md
@@ -10,6 +10,7 @@ Contents
 * Use the GtkSource provided `FileLoader` and `FileSaver` classes to do async io
 * Make sure window title gets set after `save_as` action
 * Set smart home, end and backspace key handling
+* Open files from command line
 
 ## 0.1.0
 First basic proof of concept release.